\name{plot.select}
\alias{plot.select}
\title{
Plot Function for "select"
}
\description{
Plot the optimal graph by model selection. 
}
\usage{
\method{plot}{select}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"select"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{sugm}} and \code{\link{sugm.select}}
}
