\name{plot.sugm}
\alias{plot.sugm}
\title{
Plot Function for "sugm"
}
\description{
Plot sparsity level information and 3 typical sparse graphs from the graph path.
}
\usage{
\method{plot}{sugm}(x, align = FALSE, \dots)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"sugm"} 
  }
  \item{align}{
  If \code{align = FALSE}, 3 plotted graphs are aligned
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{sugm}} and \code{\link{flare-package}}
}
