\name{print.roc}
\alias{print.roc}
\title{
Print Function for an object with S3 class \code{"roc"}
}
\description{
Print true positive rates, false positive rates, area under the curve, and maximum F1 score.
}
\usage{
\method{print}{roc}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"roc"}
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Tuo Zhao <tourzhao@gatech.edu>
}
\seealso{
\code{\link{sugm.roc}}, \code{\link{sugm}} and \code{\link{flare-package}}
}
