\name{print.select}
\alias{print.select}
\title{
Print function for S3 class "select"
}
\description{
Print the information about the model usage, graph dimension, model selection criterion, sparsity level of the optimal graph
}
\usage{
\method{print}{select}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"select"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{flare.tiger.select}} and \code{\link{flare.tiger}}
}
