\name{print.sim}
\alias{print.sim}
\title{
Print Function for an object with S3 class \code{"sim"}
}
\description{
Print the information about the sample size, the dimension, the pattern and sparsity of the true graph structure.
}
\usage{
\method{print}{sim}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"sim"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}
\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Tuo Zhao <tourzhao@gatech.edu>
}
\seealso{
\code{\link{sugm}} and \code{\link{sugm.generator}}
}
