\name{print.slim}
\alias{print.slim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print a flare.slim Object
}
\description{
Print a summary of the information about the slim object.
}
\usage{
\method{print}{slim}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  The slim object.
}
  \item{\dots}{
  Additional print options.
}
}
\details{
This call simply outlines the options used for computing a slim object.
}
%\value{
%}
%\references{
%}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{flare.slim}} and \code{\link{flare-package}}.
}
