\name{print.tiger}
\alias{print.tiger}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print a flare.tiger Object
}
\description{
Print a summary of the information about the tiger object.
}
\usage{
\method{print}{tiger}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  The tiger object.
}
  \item{\dots}{
  Additional print options.
}
}
\details{
This call simply outlines the options used for computing a tiger object.
}
%\value{
%}
%\references{
%}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{flare}} and \code{\link{flare-package}}.
}
