\name{plot.roc}
\alias{plot.roc}
\title{
Plot Function for "roc"
}
\description{
Plot the ROC curve for an object with S3 class \code{"roc"}
}
\usage{
\method{plot}{roc}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"roc"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{sugm.roc}}, \code{\link{sugm}} and \code{\link{flare-package}}.
}
