\name{plot.slim}
\alias{plot.slim}
\title{
Plot function for S3 class "slim"
}
\description{
Visualize the solution path of regression estimate corresponding to regularization paramters.
}
\usage{
\method{plot}{slim}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"slim"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{flare.slim}}
}
