\name{print.roc}
\alias{print.roc}
\title{
Print function for S3 class "roc"
}
\description{
Print the information about true positive rates, false positive rates, the area under curve and maximum F1 score
}
\usage{
\method{print}{roc}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"roc"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{flare.tiger}} and \code{\link{flare.tiger.roc}}
}
