\name{print.roc}
\alias{print.roc}
\title{
Print Function for for an object with S3 class \code{"roc"}
}
\description{
Print the information about true positive rates, false positive rates, the area under curve and maximum F1 score
}
\usage{
\method{print}{roc}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"roc"}
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{sugm.roc}}, \code{\link{sugm}} and \code{\link{flare-package}}
}
