\name{print.select}
\alias{print.select}
\title{
Print Function for an object with S3 class \code{"select"}
}
\description{
Print information about model usage, graph dimension, selection criterion, and sparsity level of the optimal graph.
}
\usage{
\method{print}{select}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"select"}
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Tuo Zhao <tourzhao@gatech.edu>
}
\seealso{
\code{\link{sugm.select}}, \code{\link{sugm}} and \code{\link{flare-package}}
}
