\name{print.select}
\alias{print.select}
\title{
Print Function for for an object with S3 class \code{"select"}
}
\description{
Print the information about the model usage, graph dimension, model selection criterion, sparsity level of the optimal graph
}
\usage{
\method{print}{select}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"select"}
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{sugm.select}}, \code{\link{sugm}} and \code{\link{flare-package}}
}
