\name{print.slim}
\alias{print.slim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Function for an object with S3 class \code{"slim"}
}
\description{
Print a summary of the information about an object with S3 class \code{"slim"}.
}
\usage{
\method{print}{slim}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  An object with S3 class \code{"slim"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}
\details{
This call simply outlines the options used for computing a slim object.
}
%\value{
%}
%\references{
%}

\author{
Xingguo Li, Tuo Zhao, Lie Wang, Xiaoming Yuan and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{slim}} and \code{\link{flare-package}}.
}
